% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group.filterNA.summary}
\alias{group.filterNA.summary}
\title{group.filterNA.summary}
\usage{
group.filterNA.summary(data, column)
}
\arguments{
\item{data}{-> this is the name of the data set}

\item{column}{-> this is the name of the column/variable of interest}
}
\value{
summary table of counts for each value
}
\description{
group.filterNA.summary
}
\examples{
group.filterNA.summary(vancouver_trees, street_side_name) #expect summary table of numbers of counts of EVEN vs. ODD
group.filterNA.summary(vancouver_trees, neighbourhood_name) #expect summary table of numbers of counts of neighbourhoods
group.filterNA.summary(vancouver_trees, curb) #expect summary table of numbers of counts of neighbourhoods
}
