% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedata.R
\docType{data}
\name{treedata}
\alias{treedata}
\title{Data of trees in Vancouver, Canada}
\format{
A tibble with 146,611 rows and 20 variables:
}
\source{
\url{https://github.com/UBC-MDS/datateachr}
}
\usage{
treedata
}
\description{
This data set includes information on the trees in the city of Vancouver. It includes 20 variables including information on tree location, surroundings, names, and more.
}
\keyword{datasets}
