% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group.filterNA.summary}
\alias{group.filterNA.summary}
\title{group.filterNA.summary}
\usage{
group.filterNA.summary(data, column)
}
\arguments{
\item{data}{-> this is the name of the data set}

\item{column}{-> this is the name of the column/variable of interest}
}
\value{
Summary table of counts for each value
}
\description{
This package takes a column and data frame input and removes all NA values from that column. Next, it creates a summary table with the count of each value in the column. This function is used when you are interested in a count summary table of values within one column of your data.
}
\examples{
group.filterNA.summary(treedata, "street_side_name") #expect summary table with EVEN vs. ODD
group.filterNA.summary(treedata, "neighbourhood_name") #expect summary of neighbourhoods
group.filterNA.summary(treedata, "curb") #expect summary of curb yes vs. no
}
\author{
Marcella Cross - marcella.cross@botany.ubc.ca - ORCID id: 0000-0002-3866-6970
}
